it remains to change the output result

+++ num_work.h
@@ -2,13 +2,13 @@
 
 #define NUM_WORK 
 struct num_list {
-    int data;
+    long data;
     int count;
     struct num_list *next;
 };
 
-struct num_list *is_num_list(int n, struct num_list *ptr);
-struct num_list *add_num_list(int n, struct num_list *ptr);
+struct num_list *is_num_list(long n, struct num_list *ptr);
+struct num_list *add_num_list(long n, struct num_list *ptr);
 void show_num_list(struct num_list *ptr);
 
 #endif
